+++ src/core/bez/index.ts
@@ -1,4 +1,27 @@
+import BezierEasing from './BezierEasing';
+
 const defaultCurveSegments = 200;
+const beziers = {};
+
+/**
+ * get a bezierEasing from real time or cache
+ * @private
+ * @param {*} a in control point x component
+ * @param {*} b in control point y component
+ * @param {*} c out control point x component
+ * @param {*} d out control point y component
+ * @param {*} [nm] curver name
+ * @return {BezierEasing}
+ */
+function getBezierEasing(a, b, c, d, nm) {
+  const str = nm || ('bez_' + a+'_'+b+'_'+c+'_'+d).replace(/\./g, 'p');
+  if (beziers[str]) {
+    return beziers[str];
+  }
+  const bezEasing = new BezierEasing(a, b, c, d);
+  beziers[str] = bezEasing;
+  return bezEasing;
+}
 
 /**
  * a
@@ -123,4 +146,5 @@ export default {
   buildBezierData,
   pointOnLine2D,
   pointOnLine3D,
+  getBezierEasing 
 };
