+++ src/core/property/BaseProperty.ts
@@ -1,5 +1,4 @@
-import BezierFactory from '../bez/BezierEaser';
-import bez from '../bez/bez';
+import bez from '../bez';
 import { Quaternion, Vector3 } from 'oasis-engine';
 import Expression from '../Expression';
 
@@ -112,7 +111,7 @@ export default class BaseProperty {
         if (keyData.__fnct) {
           fnc = keyData.__fnct;
         } else {
-          fnc = BezierFactory.getBezierEasing(keyData.o.x, keyData.o.y, keyData.i.x, keyData.i.y, keyData.n).get;
+          fnc = bez.getBezierEasing(keyData.o.x, keyData.o.y, keyData.i.x, keyData.i.y, keyData.n).get;
           keyData.__fnct = fnc;
         }
         perc = fnc((frameNum - keyTime) / (nextKeyTime - keyTime));
@@ -195,7 +194,7 @@ export default class BaseProperty {
                   outY = (typeof keyData.o.y[i] === 'undefined') ? keyData.o.y[0] : keyData.o.y[i];
                   inX = (typeof keyData.i.x[i] === 'undefined') ? keyData.i.x[0] : keyData.i.x[i];
                   inY = (typeof keyData.i.y[i] === 'undefined') ? keyData.i.y[0] : keyData.i.y[i];
-                  fnc = BezierFactory.getBezierEasing(outX, outY, inX, inY).get;
+                  fnc = bez.getBezierEasing(outX, outY, inX, inY).get;
                   keyData.__fnct[i] = fnc;
                 } else {
                   fnc = keyData.__fnct[i];
@@ -206,7 +205,7 @@ export default class BaseProperty {
                   outY = keyData.o.y;
                   inX = keyData.i.x;
                   inY = keyData.i.y;
-                  fnc = BezierFactory.getBezierEasing(outX, outY, inX, inY).get;
+                  fnc = bez.getBezierEasing(outX, outY, inX, inY).get;
                   keyData.__fnct = fnc;
                 } else {
                   fnc = keyData.__fnct;
